Fix: bad implicit type used in symmetric encryption (u32 instead of u64 for additional data)

+++ src/hybrid_crypto/block.rs
@@ -62,7 +62,10 @@ where
         // recover the block header and regenerate the IV
         let block_header = BlockHeader::<S>::parse(&encrypted_bytes[0..block_header_len])?;
         let mut ad = uid.to_vec();
-        ad.extend(&block_number.to_le_bytes());
+        // Warning: usize can be interpret as u32 on 32-bits CPU-architecture.
+        // The u64-cast prevents build on those 32-bits machine or on
+        // `wasm32-unknown-unknown` builds.
+        ad.extend(&(block_number as u64).to_le_bytes());
         // decrypt
         let clear_text = symmetric_crypto.decrypt(
             symmetric_key,
@@ -90,7 +93,10 @@ where
         // refresh the nonce
         let nonce = symmetric_crypto.generate_nonce();
         let mut ad = uid.to_vec();
-        ad.extend(&block_number.to_le_bytes());
+        // Warning: usize can be interpret as u32 on 32-bits CPU-architecture.
+        // The u64-cast prevents build on those 32-bits machine or on
+        // `wasm32-unknown-unknown` builds.
+        ad.extend(&(block_number as u64).to_le_bytes());
         // write the header
         let mut bytes = BlockHeader::<S> {
             nonce: nonce.clone(),
